gif => media component, more in line with what it can really do
